[tests] replace runners and test suite with testcase approach
