+++ tox.ini
@@ -295,7 +295,7 @@ commands =
     bottle_contrib: nosetests {posargs} tests/contrib/bottle/test.py
     bottle_contrib_autopatch: python tests/ddtrace_run.py nosetests {posargs} tests/contrib/bottle/test_autopatch.py
     cassandra_contrib: nosetests {posargs} tests/contrib/cassandra
-    celery_contrib: python -m tests.runner {posargs} tests/contrib/celery
+    celery_contrib: nosetests {posargs} tests/contrib/celery
     django_contrib: python tests/contrib/django/runtests.py {posargs}
     django_contrib_autopatch: python tests/ddtrace_run.py python tests/contrib/django/runtests.py {posargs}
     django_drf_contrib: python tests/contrib/djangorestframework/runtests.py {posargs}
@@ -306,7 +306,7 @@ commands =
     flask_contrib_autopatch: python tests/ddtrace_run.py pytest {posargs} tests/contrib/flask_autopatch
     flask_cache_contrib: nosetests {posargs} tests/contrib/flask_cache
     futures_contrib: nosetests {posargs} tests/contrib/futures
-    gevent_contrib: python -m tests.runner {posargs} tests/contrib/gevent
+    gevent_contrib: nosetests {posargs} tests/contrib/gevent
     grpc_contrib: nosetests {posargs} tests/contrib/grpc
     httplib_contrib: nosetests {posargs} tests/contrib/httplib
     jinja2_contrib: nosetests {posargs} tests/contrib/jinja2
@@ -325,7 +325,7 @@ commands =
     pyramid_contrib_autopatch: python tests/ddtrace_run.py nosetests {posargs} tests/contrib/pyramid/test_pyramid_autopatch.py
     redis_contrib: nosetests {posargs} tests/contrib/redis
     rediscluster_contrib: nosetests {posargs} tests/contrib/rediscluster
-    requests_contrib: python -m tests.runner {posargs} tests/contrib/requests
+    requests_contrib: nosetests {posargs} tests/contrib/requests
     requests_gevent_contrib: nosetests {posargs} tests/contrib/requests_gevent
     kombu_contrib: nosetests {posargs} tests/contrib/kombu
     sqlalchemy_contrib: nosetests {posargs} tests/contrib/sqlalchemy
